+++ Dockerfile
@@ -1,19 +1,19 @@
-FROM golang:1.21
+FROM golang:1.21 AS builder
 
-WORKDIR /usr/src/app 
+WORKDIR /usr/src/app
 
 COPY go.mod go.sum ./
 
 RUN go mod download
 
-COPY *.go *.db ./
-COPY ./controllers ./controllers
-COPY ./services ./services
-COPY ./nextdate ./nextdate
-COPY ./models ./models
-COPY ./web ./web
-COPY .env .
+COPY . .
+RUN CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -o /my_app ./cmd/app
 
-RUN CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -o /my_app
+FROM ubuntu:latest
 
-CMD ["/my_app"]
+WORKDIR /root/
+COPY --from=builder /my_app .
+COPY ./web ./web 
+COPY .env . 
+
+CMD ["./my_app"]
