+++ tests/settings.go
@@ -4,4 +4,4 @@ var Port = 7540
 var DBFile = "../scheduler.db"
 var FullNextDate = true
 var Search = true
-var Token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6InVzZXIiLCJleHAiOjE3MjkwNTgwMTZ9.tBTlegKVCLAMQgtYvTJLnOVQGB7m7BN05WI6uANGi6E"
+var Token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6InVzZXIiLCJleHAiOjE3Mjk2NTAyMTJ9.5PELBgz8k5FiJoqy7iQUznKIhNq1bghnGV1j3xHmv2Q"
